fix(TupleBufferRef) remove the rest of mentions of uint32_t prepended size that appeared after rebase

+++ nes-nautilus/src/Nautilus/DataTypes/VariableSizedData.cpp
@@ -104,7 +104,7 @@ nautilus::val<bool> VariableSizedData::operator!() const
     return !isValid();
 }
 
-[[nodiscard]] nautilus::val<uint32_t> VariableSizedData::getSize() const
+[[nodiscard]] nautilus::val<uint64_t> VariableSizedData::getSize() const
 {
     return size;
 }
